+++ temp_cata/gui.lua
@@ -6103,6 +6103,8 @@ local function guiFrameAuras()
         else
             changeIconButton:SetAlpha(0)
             changeIconButton:Disable()
+            BetterBlizzFramesDB.toggleIconPosition = nil
+            BBF.UpdateHiddenAuraButtonPos()
         end
     end)